간단한 수정

+++ LogicSimulator/LogicSimulator/notGate.cpp
@@ -35,15 +35,15 @@ void notGate::draw_shadow(Gdiplus::Graphics * gp, Gdiplus::Pen * p)
 
 	//Rect :: 필드상에서 표시될 위치 & 옆의 좌표는 이미지에서 잘라올 좌표
 
-	gp->DrawImage(pBitmap, Gdiplus::Rect(outputCoord.x - 3 * UNIT, outputCoord.y -1.5 * UNIT, 30, 30), 30 * 0, 60 * 5, 30, 30, Gdiplus::UnitPixel, &imAtt, NULL, NULL);
+	gp->DrawImage(pBitmap, Gdiplus::Rect(outputCoord.x - 3 * UNIT, (int)(outputCoord.y -1.5 * UNIT), 30, 30), 30 * 0, 60 * 5, 30, 30, Gdiplus::UnitPixel, &imAtt, NULL, NULL);
 
 	delete pBitmap;
 }
 
 void notGate::set_Coord_From_outC(int x, int y)
 {
-	top.x = x - 3 * UNIT; top.y = y - 1.5 * UNIT;
-	bottom.x = x; bottom.y = y + 1.5 * UNIT;
+	top.x = x - 3 * UNIT; top.y = y - (int)(1.5 * UNIT);
+	bottom.x = x; bottom.y = y + (int)(1.5 * UNIT);
 }
 
 void notGate::turn()
